requirements.txt
python-dotenv

langchain
langchain-openai
langchain-ollama
langchain-community

langchain-text-splitters
langchain-chroma
unstructured[docx,pdf]